fix: 32-bit targets cast guest addresses and wasm memory limits

+++ src/core/platform.zig
@@ -10,7 +10,7 @@ pub const ptr_bits = std.Target.ptrBitWidth(&builtin.target);
 pub const is_64bit = ptr_bits == 64;
 pub const is_32bit = ptr_bits == 32;
 
-pub const max_wasm32_linear_memory_bytes: usize = 0x1_0000_0000;
+pub const max_wasm32_linear_memory_bytes: u64 = 0x1_0000_0000;
 pub const max_linear_memory_bytes: usize = if (is_64bit)
     1 << 48 // practical host limit for memory64 linear memory
 else
